+++ src/sentry/rules/conditions/event_frequency.py
@@ -22,7 +22,6 @@ standard_intervals = {
     "30d": ("30 days", timedelta(days=30)),
 }
 comparison_intervals = {
-    "1m": ("one minute", timedelta(minutes=1)),
     "5m": ("5 minutes", timedelta(minutes=5)),
     "15m": ("15 minutes", timedelta(minutes=15)),
     "1h": ("one hour", timedelta(hours=1)),
